// music-client/src/enums/router-name.ts
export const enum RouterName {
  Home = "/",
  SongSheet = "/song-sheet",
  SongSheetDetail = "/song-sheet-detail",
  Singer = "/singer",
  SingerDetail = "/singer-detail",
  Community = "/community",
  Social = "/social",
  CommunityPublish = "/community/publish",
  CommunityDetail = "/community/detail",
  Personal = "/personal",
  Setting = "/setting",
  Admin = "/admin",
  AdminDashboard = "/admin/dashboard",
  AdminSong = "/admin/song",
  AdminUser = "/admin/user",
  AdminComment = "/admin/comment",
  AdminSongList = "/admin/song-list",
  AdminPostAudit = "/admin/post-audit",
  PersonalData = "/personal-data",
  SignIn = "/sign-in",
  SignUp = "/sign-up",
  Search = "/search",
  Lyric = "/lyric",
  Error = "/404",
  SignOut = "0",
}